relocate the util because it is not the feature of the previous package
